requirements.txt
certifi==2021.5.30
click==8.0.1
cycler==0.10.0
dataclasses==0.8
Flask==1.1.2
future==0.18.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.17.0
nvidia-dali-cuda100==1.3.0
onnx==1.6.0
onnxruntime==1.3.0
opencv-python==4.2.0.34
parse==1.19.0
Pillow==8.3.2
protobuf==3.18.0
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
torch==1.6.0
torchvision==0.7.0
typing-extensions==3.10.0.2
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.5.0
